lightining/requirements.txt
# # --------- pytorch --------- #
--extra-index-url https://download.pytorch.org/whl/cu117
torch==2.0.0+cu117
torchvision==0.15.1+cu117

lightning[extra]==2.0.0
torchmetrics==0.11.4

# # # --------- model zoo -------#
transformers==4.30.2
timm==0.9.2

# # ------- tensor utils ------#
einops==0.6.1

# # ------- loggers ------#
mlflow==2.4.1
aim==3.17.5

# # --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0
hydra-joblib-launcher==1.2.0

# # --------- others --------- #
pyrootutils==1.0.4      # standardizing the project root setup
pre-commit==3.3.3       # hooks for applying linters on commit
rich==13.4.2            # beautiful text formatting in terminal
pytest==7.3.2           # testing
black==23.3.0           # code formatting
twine==4.0.2            # for uploading to pypi
sh==2.0.4               # for running bash commands in some tests (linux/macos only)
tiktoken==0.4.0         # for openai embeddings        
gradio==3.39.0          # for creating demo web apps

# # --------- data version control --------- #
dvc==3.0.0              # data version control
dvc-gdrive==2.20.0      # google drive integration